Moved compatibility checks to header and internal namespace.
