+++ ros_babel_fish/src/message.cpp
@@ -1,12 +1,12 @@
 // Copyright (c) 2019 Stefan Fabian. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+#include "ros_babel_fish/message.h"
+#include "ros_babel_fish/messages/internal/value_compatibility.h"
 #include "ros_babel_fish/messages/compound_message.h"
 #include "ros_babel_fish/messages/value_message.h"
 #include "ros_babel_fish/babel_fish.h"
-#include "ros_babel_fish/message.h"
 
-#include <limits>
 
 namespace ros_babel_fish
 {
@@ -30,84 +30,13 @@ Message::~Message() = default;
 namespace
 {
 
-// Backport of C++20's std::remove_cvref
-template<class T>
-struct rm_cvref
-{
-  typedef typename std::remove_cv<typename std::remove_reference<T>::type>::type type;
-};
-template<class T>
-using rm_cvref_t = typename rm_cvref<T>::type;
-
-// is_integral is necessary to avoid a CLang tidy warning
-template<typename T, typename U>
-typename std::enable_if<
-  std::is_integral<T>::value &&
-  std::numeric_limits<T>::is_signed && !std::numeric_limits<U>::is_signed, bool>::type
-constexpr inBounds( const T &val )
-{
-  return val >= 0 && static_cast<typename std::make_unsigned<T>::type >( val ) <= std::numeric_limits<U>::max();
-}
-
-template<typename T, typename U>
-typename std::enable_if<
-  std::is_integral<T>::value &&
-  !std::numeric_limits<T>::is_signed && std::numeric_limits<U>::is_signed &&
-  std::numeric_limits<T>::digits >= std::numeric_limits<U>::digits, bool>::type
-constexpr inBounds( const T &val )
-{
-  return val <= static_cast<T>(std::numeric_limits<U>::max());
-}
-
-template<typename T, typename U>
-typename std::enable_if<
-  std::is_integral<T>::value &&
-  !std::numeric_limits<T>::is_signed && std::numeric_limits<U>::is_signed &&
-  std::numeric_limits<T>::digits < std::numeric_limits<U>::digits, bool>::type
-constexpr inBounds( const T &val )
-{
-  return static_cast<U>(val) <= std::numeric_limits<U>::max();
-}
-
-template<typename T, typename U>
-typename std::enable_if<
-  std::is_integral<T>::value &&
-  std::numeric_limits<T>::is_signed == std::numeric_limits<U>::is_signed, bool>::type
-constexpr inBounds( const T &val )
-{
-  return std::numeric_limits<U>::min() <= val && val <= std::numeric_limits<U>::max();
-}
-
-template<typename T, typename U>
-typename std::enable_if<std::is_floating_point<T>::value, bool>::type
-constexpr inBounds( const T &val )
-{
-  return static_cast<T>(std::numeric_limits<U>::min()) <= val && val <= static_cast<T>(std::numeric_limits<U>::max());
-}
-
-/**
- * Returns true if type T can always be stored in U without overflowing.
- */
-template<typename T, typename U>
-typename std::enable_if<std::is_arithmetic<T>::value && std::is_arithmetic<U>::value, bool>::type
-constexpr isCompatible()
-{
-  // See https://en.cppreference.com/w/cpp/types/numeric_limits/digits
-  // + 1 added because ::digits returns nbits-1 for signed types.
-  return std::is_same<rm_cvref_t<T>, rm_cvref_t<U>>::value ||
-         std::is_floating_point<U>::value ||
-         (std::is_integral<T>::value &&
-          !(std::is_signed<T>::value && std::is_unsigned<U>::value) &&
-          (std::numeric_limits<T>::digits + 1 < std::numeric_limits<U>::digits));
-}
-
 template<typename T, typename U>
 void assignValue( Message *m, const T &value )
 {
   using namespace message_type_traits;
-  if ( !isCompatible<T, U>())
+  if ( !internal::isCompatible<T, U>())
   {
-    if ( !inBounds<T, U>( value ))
+    if ( !internal::inBounds<T, U>( value ))
     {
       throw BabelFishException(
         "Value does not fit into value message! Make sure you're using the correct type or at least stay within the range of values for the message type!" );
@@ -270,9 +199,9 @@ U obtainValue( const Message *m )
 {
   using namespace message_type_traits;
   T val = m->as<ValueMessage<T>>().getValue();
-  if ( !isCompatible<T, U>())
+  if ( !internal::isCompatible<T, U>())
   {
-    if ( !inBounds<T, U>( val ))
+    if ( !internal::inBounds<T, U>( val ))
     {
       throw BabelFishException( "Value does not fit into casted type!" );
     }
